+++ changelog.md
@@ -1,5 +1,6 @@
 # Changelog
 ## v0.4.7
+- Fixed exception when try to insert in horus_entities with a entity name exists.
 - Fixed issue in Entity get attributes polymorphic like getInt and getFloat.
 
 ## v0.4.6